% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{isoexplorer-package}
\alias{isoexplorer}
\alias{isoexplorer-package}
\title{isoexplorer: GUI to Explore Stable Isotope Data Files}
\description{
Provides a graphical user interface to explore stable isotope data files using the isoreader2 package, including browsing, visualizing, and exporting isotope data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/isoverse/isoexplorer}
  \item Report bugs at \url{https://github.com/isoverse/isoexplorer/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{seb.kopf@gmail.com}

Authors:
\itemize{
  \item Sebastian Kopf \email{seb.kopf@gmail.com}
}

}
\keyword{internal}
